Project card

+++ src/main/java/com/example/naumenProject/controllers/WebController.java
@@ -62,6 +62,21 @@ public class WebController {
         return "index";
     }
 
+    @GetMapping(value = "/all_projects")
+    public String getAllProjectsPage(Model model, Authentication authentication) {
+        String username = authentication.getName();
+
+        User currentUser = userRepository.findUserByUsername(username);
+
+        var projects = projectService.getAllProjects();
+
+        model.addAttribute("title", "Проекты");
+        model.addAttribute("user", currentUser);
+        model.addAttribute("projects", projects);
+
+        return "projects";
+    }
+
     @GetMapping(value = "/projects")
     public String getProjectsPage(Model model, Authentication authentication) {
         String username = authentication.getName();
@@ -70,7 +85,7 @@ public class WebController {
 
         var projects = projectService.getAllProjects();
 
-        model.addAttribute("title", "Проекты");
+        model.addAttribute("title", "Мои проекты");
         model.addAttribute("user", currentUser);
         model.addAttribute("projects", projects);
 
@@ -107,25 +122,14 @@ public class WebController {
         return "redirect:/projects";
     }
 
-//    @PostMapping(value = "/updateProject")
-//    public String updateProject(@RequestParam("name") String name, @RequestParam("description") String description,
-//                                Authentication authentication) {
-//        String username = authentication.getName();
-//
-//        User currentUser = userRepository.findUserByUsername(username);
-//
-//        Project project = new Project(UUID.randomUUID().getMostSignificantBits(), username, name, description, "", "", "", "",
-//                "");
-//
-//        projectService.createProject(project);
-//
-//        return "redirect:/projects";
-//    }
-
-    // Не работает, ПОПРАВИТЬ удаление только TeamLead
     @PostMapping(value = "/deleteProject")
     public String deleteProject(@RequestParam("id") Long id, Authentication authentication) {
-        projectService.deleteProject(id);
+        String username = authentication.getName();
+        Project project = projectService.getProjectById(id);
+        String creatorName = project.getProjectCreator();
+        if (Objects.equals(username, creatorName)){
+            projectService.deleteProject(id);
+        }
 
         return "redirect:/projects";
     }
@@ -368,5 +372,20 @@ public class WebController {
         } else {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid project ID or empty file.");
         }
+    }
+
+    @PostMapping(value = "/chooseTeamRole")
+    public String chooseTeamRole(@RequestParam("team_role") String roleInProject,
+                                 Authentication authentication) {
+
+        String username = authentication.getName();
+        var user = userService.getUserByUsername(username);
+
+        if (user != null) {
+            user.setRoleInProject(roleInProject);
+            userService.createUser(user);
+        }
+
+        return "redirect:/";
     }
 }
